requirements.txt
Flask==3.0.0
chromadb==0.4.22
sentence-transformers==2.3.1
PyMuPDF==1.23.8
python-dotenv==1.0.0
google-generativeai==0.3.2
python-docx==1.1.0
Werkzeug==3.0.1